+++ app/javascript/src/components/Blog/Lists/List.jsx
