+++ app/javascript/src/components/Blog/constants.js
@@ -1,6 +1,10 @@
 import { t } from "i18next";
 import * as yup from "yup";
 
+export const DEFAULT_PAGE_INDEX = 1;
+
+export const DEFAULT_PAGE_SIZE = 5;
+
 export const NEW_POST_INITIAL_VALUES = {
   title: "",
   description: "",
