--- app/javascript/src/components/BlogLists/constants.js
@@ -1,3 +1,0 @@
-export const DEFAULT_PAGE_INDEX = 1;
-
-export const DEFAULT_PAGE_SIZE = 5;
