+++ app/views/posts/index.json.jbuilder
@@ -14,4 +14,4 @@ json.posts @posts do |post|
   end
 end
 
-json.total_posts_count @posts.count
+json.total_posts_count @total_posts_count
